schedule-microservice/requirements.txt
Django~=5.0
requests
djangorestframework~=3.0
psycopg2-binary~=2.0
djangorestframework_simplejwt
django-environ
bcrypt
pika==1.3.1
django-cors-headers